requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-login
python-dotenv
psycopg2-binary
gunicorn
marshmallow
